Funcionando Address Interfaz y Service, arrreglar editar y eliminar de determinados dispositivos

+++ smartnetwork-backend/src/main/java/com/smartnetwork/backend/Repository/AddressRepository.java
@@ -9,6 +9,6 @@ import java.util.Optional;
 
 @Repository
 public interface AddressRepository extends JpaRepository<Address,Long> {
-    List<Address> findByUsuarioId(Integer id);
+    List<Address> findByUsuarioId(Long id);
 
 }
